+++ packages/block-library/src/cover/index.js
@@ -214,6 +214,7 @@ export const settings = {
 					}
 					mediaType = media.type;
 				}
+
 				setAttributes( {
 					url: media.url,
 					id: media.id,