+++ packages/editor/src/components/navigable-toolbar/index.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { cond, matchesProperty } from 'lodash';
+import { cond, matchesProperty, omit } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -75,7 +75,9 @@ class NavigableToolbar extends Component {
 				role="toolbar"
 				ref={ this.toolbar }
 				onKeyDown={ this.switchOnKeyDown }
-				{ ...props }
+				{ ...omit( props, [
+					'focusOnMount',
+				] ) }
 			>
 				<KeyboardShortcuts
 					bindGlobal