+++ packages/editor/src/components/rich-text/tinymce.js
@@ -175,7 +175,8 @@ export default class TinyMCE extends Component {
 	}
 
 	initialize() {
-		const settings = this.props.getSettings( {
+		const { multilineTag } = this.props;
+		const settings = {
 			theme: false,
 			inline: true,
 			toolbar: false,
@@ -189,7 +190,16 @@ export default class TinyMCE extends Component {
 			verify_html: false,
 			inline_boundaries_selector: 'a[href],code,b,i,strong,em,del,ins,sup,sub',
 			plugins: [],
-		} );
+			forced_root_block: multilineTag || false,
+			// Allow TinyMCE to keep one undo level for comparing changes.
+			// Prevent it otherwise from accumulating any history.
+			custom_undo_redo_levels: 1,
+			lists_indent_on_tab: false,
+		};
+
+		if ( multilineTag === 'li' ) {
+			settings.plugins.push( 'lists' );
+		}
 
 		tinymce.init( {
 			...settings,
@@ -211,7 +221,18 @@ export default class TinyMCE extends Component {
 				} );
 
 				editor.on( 'init', () => {
-					// See https://github.com/tinymce/tinymce/blob/master/src/core/main/ts/keyboard/FormatShortcuts.ts
+					// History is handled internally by RichText.
+					//
+					// See: https://github.com/tinymce/tinymce/blob/master/src/core/main/ts/api/UndoManager.ts
+					[ 'z', 'y' ].forEach( ( character ) => {
+						editor.shortcuts.remove( `meta+${ character }` );
+					} );
+					editor.shortcuts.remove( 'meta+shift+z' );
+
+					// Reset TinyMCE's default formatting shortcuts, since
+					// RichText supports only registered formats.
+					//
+					// See: https://github.com/tinymce/tinymce/blob/master/src/core/main/ts/keyboard/FormatShortcuts.ts
 					[ 'b', 'i', 'u' ].forEach( ( character ) => {
 						editor.shortcuts.remove( `meta+${ character }` );
 					} );
@@ -219,6 +240,7 @@ export default class TinyMCE extends Component {
 						editor.shortcuts.remove( `access+${ number }` );
 					} );
 
+					// Restore the original `setHTML` once initialized.
 					editor.dom.setHTML = setHTML;
 				} );
 
@@ -315,6 +337,7 @@ export default class TinyMCE extends Component {
 			onInput,
 			onKeyDown,
 			onKeyUp,
+			onCompositionEnd,
 		} = this.props;
 
 		/*
@@ -344,6 +367,7 @@ export default class TinyMCE extends Component {
 			onFocus: this.onFocus,
 			onKeyDown,
 			onKeyUp,
+			onCompositionEnd,
 		} );
 	}
 }