+++ packages/editor/src/components/url-input/index.js
@@ -9,7 +9,7 @@ import scrollIntoView from 'dom-scroll-into-view';
  * WordPress dependencies
  */
 import { __, sprintf, _n } from '@wordpress/i18n';
-import { Component, Fragment, createRef } from '@wordpress/element';
+import { Component, createRef } from '@wordpress/element';
 import { decodeEntities } from '@wordpress/html-entities';
 import { UP, DOWN, ENTER, TAB } from '@wordpress/keycodes';
 import { Spinner, withSpokenMessages, Popover } from '@wordpress/components';
@@ -138,6 +138,38 @@ class URLInput extends Component {
 		// If the suggestions are not shown or loading, we shouldn't handle the arrow keys
 		// We shouldn't preventDefault to allow block arrow keys navigation
 		if ( ! showSuggestions || ! posts.length || loading ) {
+			// In the Windows version of Firefox the up and down arrows don't move the caret
+			// within an input field like they do for Mac Firefox/Chrome/Safari. This causes
+			// a form of focus trapping that is disruptive to the user experience. This disruption
+			// only happens if the caret is not in the first or last position in the text input.
+			// See: https://github.com/WordPress/gutenberg/issues/5693#issuecomment-436684747
+			switch ( event.keyCode ) {
+				// When UP is pressed, if the caret is at the start of the text, move it to the 0
+				// position.
+				case UP: {
+					if ( 0 !== event.target.selectionStart ) {
+						event.stopPropagation();
+						event.preventDefault();
+
+						// Set the input caret to position 0
+						event.target.setSelectionRange( 0, 0 );
+					}
+					break;
+				}
+				// When DOWN is pressed, if the caret is not at the end of the text, move it to the
+				// last position.
+				case DOWN: {
+					if ( this.props.value.length !== event.target.selectionStart ) {
+						event.stopPropagation();
+						event.preventDefault();
+
+						// Set the input caret to the last position
+						event.target.setSelectionRange( this.props.value.length, this.props.value.length );
+					}
+					break;
+				}
+			}
+
 			return;
 		}
 
@@ -199,28 +231,26 @@ class URLInput extends Component {
 		const { showSuggestions, posts, selectedSuggestion, loading } = this.state;
 		/* eslint-disable jsx-a11y/no-autofocus */
 		return (
-			<Fragment>
-				<div className="editor-url-input">
-					<input
-						autoFocus={ autoFocus }
-						type="text"
-						aria-label={ __( 'URL' ) }
-						required
-						value={ value }
-						onChange={ this.onChange }
-						onInput={ stopEventPropagation }
-						placeholder={ __( 'Paste URL or type to search' ) }
-						onKeyDown={ this.onKeyDown }
-						role="combobox"
-						aria-expanded={ showSuggestions }
-						aria-autocomplete="list"
-						aria-owns={ `editor-url-input-suggestions-${ instanceId }` }
-						aria-activedescendant={ selectedSuggestion !== null ? `editor-url-input-suggestion-${ instanceId }-${ selectedSuggestion }` : undefined }
-						ref={ this.inputRef }
-					/>
+			<div className="editor-url-input">
+				<input
+					autoFocus={ autoFocus }
+					type="text"
+					aria-label={ __( 'URL' ) }
+					required
+					value={ value }
+					onChange={ this.onChange }
+					onInput={ stopEventPropagation }
+					placeholder={ __( 'Paste URL or type to search' ) }
+					onKeyDown={ this.onKeyDown }
+					role="combobox"
+					aria-expanded={ showSuggestions }
+					aria-autocomplete="list"
+					aria-owns={ `editor-url-input-suggestions-${ instanceId }` }
+					aria-activedescendant={ selectedSuggestion !== null ? `editor-url-input-suggestion-${ instanceId }-${ selectedSuggestion }` : undefined }
+					ref={ this.inputRef }
+				/>
 
-					{ ( loading ) && <Spinner /> }
-				</div>
+				{ ( loading ) && <Spinner /> }
 
 				{ showSuggestions && !! posts.length &&
 					<Popover position="bottom" noArrow focusOnMount={ false }>
@@ -249,7 +279,7 @@ class URLInput extends Component {
 						</div>
 					</Popover>
 				}
-			</Fragment>
+			</div>
 		);
 		/* eslint-enable jsx-a11y/no-autofocus */
 	}